extern int option
