+++ SD1_Audrius_Ivko_PIT-21-NT-I/Header.h
@@ -1,14 +1,19 @@
-
-#include <iomanip>
+#include <list>
 #include <cmath>
 #include <string>
 #include <vector>
+#include <string>
+#include <random>
+#include <chrono>
+#include <time.h>
+#include <istream>
+#include <iomanip>
+#include <sstream>
 #include <fstream>
+#include <iostream>
 #include <stdlib.h>
 #include <algorithm>
-#include <iostream>
-#include <istream>
-
+#include <exception>
 
 using	std::cout;
 using	std::cin;
@@ -23,3 +28,8 @@ using	std::vector;
 using	std::ostream;
 using	std::istream;
 using	std::sort;
+using	std::getline;
+using	std::ifstream;
+using	std::istringstream;
+using	std::to_string;
+using	std::runtime_error;
