+++ SD1_Audrius_Ivko_PIT-21-NT-I/global.h
@@ -1,0 +1,3 @@
+#include "Header.h"
+
+extern int option;
